% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{muiselib_setup}
\alias{muiselib_setup}
\title{Initialize muiselib defaults}
\usage{
muiselib_setup()
}
\description{
This function sets the global ggplot2 theme and default color palettes
to the muiselib standard.
}
